Рефакторинг + покрытие тестами

+++ loans-service/src/main/java/com/synloans/loans/service/syndicate/SyndicateParticipantService.java
@@ -1,10 +1,12 @@
 package com.synloans.loans.service.syndicate;
 
 import com.synloans.loans.model.entity.company.Bank;
+import com.synloans.loans.model.entity.loan.Loan;
 import com.synloans.loans.model.entity.loan.LoanRequest;
 import com.synloans.loans.model.entity.syndicate.Syndicate;
 import com.synloans.loans.model.entity.syndicate.SyndicateParticipant;
 import com.synloans.loans.repository.syndicate.SyndicateParticipantRepository;
+import com.synloans.loans.service.exception.SyndicateQuitException;
 import com.synloans.loans.service.exception.notfound.LoanRequestNotFoundException;
 import com.synloans.loans.service.loan.LoanRequestService;
 import lombok.RequiredArgsConstructor;
@@ -42,6 +44,10 @@ public class SyndicateParticipantService {
     public void quitFromSyndicate(Long loanRequestId, Bank bank) {
         for (SyndicateParticipant syndicateParticipant: bank.getSyndicates()){
             if (Objects.equals(syndicateParticipant.getSyndicate().getRequest().getId(), loanRequestId)){
+                Loan loan = syndicateParticipant.getSyndicate().getRequest().getLoan();
+                if (loan != null){
+                    throw new SyndicateQuitException("Нельзя выйти из синдиката, когда кредит уже выдан");
+                }
                 participantRepository.delete(syndicateParticipant);
                 return;
             }
